+++ spec/02-integration/05-proxy/10-balancer/01-healthchecks_spec.lua
@@ -493,6 +493,8 @@ describe("workspace-" .. workspace, function ()
       -- remove the upstream
       if strategy ~= "off" then
         bu.remove_upstream(bp, upstream_id)
+      else
+        bp.done()
       end
 
       -- add the upstream again